# Go/client-code-main/api/swagger.yaml
openapi: 3.0.0
info:
  title: 3gpp-monitoring-event
  description: |
    API for Monitoring Event.
    Â© 2020, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).
    All rights reserved.
  version: 1.1.0
...
servers:
- url: "{apiRoot}/3gpp-monitoring-event/v1"
  variables:
    apiRoot:
      description: apiRoot as defined in subclause 5.2.4 of 3GPP TS 29.122.
      default: https://example.com
security:
- {}
- oAuth2ClientCredentials: []
paths:
  /{scsAsId}/subscriptions:
    get:
      tags:
      - MonitoringEvent API SCS/AS level GET Operation
      summary: read all of the active subscriptions for the SCS/AS
      parameters:
      - name: scsAsId
        in: path
        description: Identifier of the SCS/AS
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK (Successful get all of the active subscriptions for the
            SCS/AS)
          content:
            application/json:
              schema:
                minItems: 0
                type: array
                description: Monitoring event subscriptions
                items:
                  $ref: '#/components/schemas/MonitoringEventSubscription'
                x-content-type: application/json
        "400":
          $ref: ./TS29122_CommonData.yaml#/components/responses/400
        "401":
          $ref: ./TS29122_CommonData.yaml#/components/responses/401
        "403":
          $ref: ./TS29122_CommonData.yaml#/components/responses/403
        "404":
          $ref: ./TS29122_CommonData.yaml#/components/responses/404
        "406":
          $ref: ./TS29122_CommonData.yaml#/components/responses/406
        "429":
          $ref: ./TS29122_CommonData.yaml#/components/responses/429
        "500":
          $ref: ./TS29122_CommonData.yaml#/components/responses/500
        "503":
          $ref: ./TS29122_CommonData.yaml#/components/responses/503
        default:
          $ref: ./TS29122_CommonData.yaml#/components/responses/default
    post:
      tags:
      - MonitoringEvent API Subscription level POST Operation
      summary: Creates a new subscription resource for monitoring event notification
      parameters:
      - name: scsAsId
        in: path
        description: Identifier of the SCS/AS
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Subscription for notification about monitoring event
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitoringEventSubscription'
        required: true
      responses:
        "200":
          description: The operation is successful and immediate report is included.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringEventReport'
        "201":
          description: Created (Successful creation of subscription)
          headers:
            Location:
              description: Contains the URI of the newly created resource
              required: true
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringEventSubscription'
        "400":
          $ref: ./TS29122_CommonData.yaml#/components/responses/400
        "401":
          $ref: ./TS29122_CommonData.yaml#/components/responses/401
        "403":
          $ref: ./TS29122_CommonData.yaml#/components/responses/403
        "404":
          $ref: ./TS29122_CommonData.yaml#/components/responses/404
        "411":
          $ref: ./TS29122_CommonData.yaml#/components/responses/411
        "413":
          $ref: ./TS29122_CommonData.yaml#/components/responses/413
        "415":
          $ref: ./TS29122_CommonData.yaml#/components/responses/415
        "429":
          $ref: ./TS29122_CommonData.yaml#/components/responses/429
        "500":
          $ref: ./TS29122_CommonData.yaml#/components/responses/500
        "503":
          $ref: ./TS29122_CommonData.yaml#/components/responses/503
        default:
          $ref: ./TS29122_CommonData.yaml#/components/responses/default
      callbacks:
        notificationDestination:
          '{request.body#/notificationDestination}':
            post:
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/MonitoringNotification'
                required: true
              responses:
                "204":
                  description: No Content (successful notification)
                "400":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/400
                "401":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/401
                "403":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/403
                "404":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/404
                "411":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/411
                "413":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/413
                "415":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/415
                "429":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/429
                "500":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/500
                "503":
                  $ref: ./TS29122_CommonData.yaml#/components/responses/503
                default:
                  $ref: ./TS29122_CommonData.yaml#/components/responses/default
  /{scsAsId}/subscriptions/{subscriptionId}:
    get:
      tags:
      - MonitoringEvent API Subscription level GET Operation
      summary: read an active subscriptions for the SCS/AS and the subscription Id
      parameters:
      - name: scsAsId
        in: path
        description: Identifier of the SCS/AS
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: subscriptionId
        in: path
        description: Identifier of the subscription resource
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK (Successful get the active subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringEventSubscription'
        "400":
          $ref: ./TS29122_CommonData.yaml#/components/responses/400
        "401":
          $ref: ./TS29122_CommonData.yaml#/components/responses/401
        "403":
          $ref: ./TS29122_CommonData.yaml#/components/responses/403
        "404":
          $ref: ./TS29122_CommonData.yaml#/components/responses/404
        "406":
          $ref: ./TS29122_CommonData.yaml#/components/responses/406
        "429":
          $ref: ./TS29122_CommonData.yaml#/components/responses/429
        "500":
          $ref: ./TS29122_CommonData.yaml#/components/responses/500
        "503":
          $ref: ./TS29122_CommonData.yaml#/components/responses/503
        default:
          $ref: ./TS29122_CommonData.yaml#/components/responses/default
    put:
      tags:
      - MonitoringEvent API subscription level PUT Operation
      summary: Updates/replaces an existing subscription resource
      parameters:
      - name: scsAsId
        in: path
        description: Identifier of the SCS/AS
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: subscriptionId
        in: path
        description: Identifier of the subscription resource
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Parameters to update/replace the existing subscription
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitoringEventSubscription'
        required: true
      responses:
        "200":
          description: OK (Successful update of the subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringEventSubscription'
        "400":
          $ref: ./TS29122_CommonData.yaml#/components/responses/400
        "401":
          $ref: ./TS29122_CommonData.yaml#/components/responses/401
        "403":
          $ref: ./TS29122_CommonData.yaml#/components/responses/403
        "404":
          $ref: ./TS29122_CommonData.yaml#/components/responses/404
        "411":
          $ref: ./TS29122_CommonData.yaml#/components/responses/411
        "413":
          $ref: ./TS29122_CommonData.yaml#/components/responses/413
        "415":
          $ref: ./TS29122_CommonData.yaml#/components/responses/415
        "429":
          $ref: ./TS29122_CommonData.yaml#/components/responses/429
        "500":
          $ref: ./TS29122_CommonData.yaml#/components/responses/500
        "503":
          $ref: ./TS29122_CommonData.yaml#/components/responses/503
        default:
          $ref: ./TS29122_CommonData.yaml#/components/responses/default
    delete:
      tags:
      - MonitoringEvent API Subscription level DELETE Operation
      summary: Deletes an already existing monitoring event subscription
      parameters:
      - name: scsAsId
        in: path
        description: Identifier of the SCS/AS
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: subscriptionId
        in: path
        description: Identifier of the subscription resource
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK (Successful deletion of the existing subscription)
          content:
            application/json:
              schema:
                minItems: 1
                type: array
                description: "The subscription was terminated successfully, the monitoring\
                  \ event report(s) shall be included if received."
                items:
                  $ref: '#/components/schemas/MonitoringEventReport'
                x-content-type: application/json
        "204":
          description: No Content (Successful deletion of the existing subscription)
        "400":
          $ref: ./TS29122_CommonData.yaml#/components/responses/400
        "401":
          $ref: ./TS29122_CommonData.yaml#/components/responses/401
        "403":
          $ref: ./TS29122_CommonData.yaml#/components/responses/403
        "404":
          $ref: ./TS29122_CommonData.yaml#/components/responses/404
        "429":
          $ref: ./TS29122_CommonData.yaml#/components/responses/429
        "500":
          $ref: ./TS29122_CommonData.yaml#/components/responses/500
        "503":
          $ref: ./TS29122_CommonData.yaml#/components/responses/503
        default:
          $ref: ./TS29122_CommonData.yaml#/components/responses/default
components:
  schemas:
    MonitoringEventSubscription:
      required:
      - monitoringType
      - notificationDestination
      type: object
      properties:
        self:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/Link
        supportedFeatures:
          $ref: ./TS29571_CommonData.yaml#/components/schemas/SupportedFeatures
        mtcProviderId:
          type: string
          description: Identifies the MTC Service Provider and/or MTC Application.
        externalId:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/ExternalId
        msisdn:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/Msisdn
        externalGroupId:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/ExternalGroupId
        addExtGroupId:
          minItems: 2
          type: array
          items:
            $ref: ./TS29122_CommonData.yaml#/components/schemas/ExternalGroupId
        ipv4Addr:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/Ipv4Addr
        ipv6Addr:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/Ipv6Addr
        notificationDestination:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/Link
        requestTestNotification:
          type: boolean
          description: Set to true by the SCS/AS to request the SCEF to send a test
            notification as defined in subclauseÂ 5.2.5.3. Set to false or omitted
            otherwise.
        websockNotifConfig:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/WebsockNotifConfig
        monitoringType:
          $ref: '#/components/schemas/MonitoringType'
        maximumNumberOfReports:
          minimum: 1
          type: integer
          description: "Identifies the maximum number of event reports to be generated\
            \ by the HSS, MME/SGSN as specified in subclauseÂ 5.6.0 of 3GPPÂ TSÂ 23.682Â\
            \ [2]."
        monitorExpireTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DateTime
        repPeriod:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        groupReportGuardTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        maximumDetectionTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        reachabilityType:
          $ref: '#/components/schemas/ReachabilityType'
        maximumLatency:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        maximumResponseTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        suggestedNumberOfDlPackets:
          minimum: 0
          type: integer
          description: "If \"monitoringType\" is \"UE_REACHABILITY\", this parameter\
            \ may be included to identify the number of packets that the serving gateway\
            \ shall buffer in case that the UE is not reachable."
        idleStatusIndication:
          type: boolean
          description: "If \"monitoringType\" is set to \"UE_REACHABILITY\" or \"\
            AVAILABILITY_AFTER_DDN_FAILURE\", this parameter may be included to indicate\
            \ the notification of when a UE, for which PSM is enabled, transitions\
            \ into idle mode. -\t\"true\"  indicate enabling of notification -\t\"\
            false\"  indicate no need to notify Default  \"false\"."
        locationType:
          $ref: '#/components/schemas/LocationType'
        accuracy:
          $ref: '#/components/schemas/Accuracy'
        minimumReportInterval:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        maxRptExpireIntvl:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        samplingInterval:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        reportingLocEstInd:
          type: boolean
          description: Indicates whether to request the location estimate for event
            reporting.
        linearDistance:
          $ref: ./TS29572_Nlmf_Location.yaml#/components/schemas/LinearDistance
        locQoS:
          $ref: ./TS29572_Nlmf_Location.yaml#/components/schemas/LocationQoS
        svcId:
          $ref: ./TS29515_Ngmlc_Location.yaml#/components/schemas/ServiceIdentity
        ldrType:
          $ref: ./TS29572_Nlmf_Location.yaml#/components/schemas/LdrType
        velocityRequested:
          $ref: ./TS29572_Nlmf_Location.yaml#/components/schemas/VelocityRequested
        maxAgeOfLocEst:
          $ref: ./TS29572_Nlmf_Location.yaml#/components/schemas/AgeOfLocationEstimate
        locTimeWindow:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/TimeWindow
        associationType:
          $ref: '#/components/schemas/AssociationType'
        plmnIndication:
          type: boolean
          description: "If \"monitoringType\" is \"ROAMING_STATUS\", this parameter\
            \ may be included to indicate the notification of UE's Serving PLMN ID.\
            \ -\t\"true\"  The value shall be used to indicate enabling of notification;\
            \ -\t\"false\"  The value shall be used to indicate disabling of notification.\
            \  Default  \"false\"."
        locationArea:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/LocationArea
        locationArea5G:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/LocationArea5G
        dddTraDescriptors:
          minItems: 1
          type: array
          items:
            $ref: ./TS29571_CommonData.yaml#/components/schemas/DddTrafficDescriptor
        dddStati:
          minItems: 1
          type: array
          items:
            $ref: ./TS29571_CommonData.yaml#/components/schemas/DlDataDeliveryStatus
        apiNames:
          minItems: 1
          type: array
          items:
            type: string
        monitoringEventReport:
          $ref: '#/components/schemas/MonitoringEventReport'
      example:
        notificationDestination: null
        locQoS: "{}"
        dddStati:
        - "{}"
        - "{}"
        monitoringEventReport:
          locationInfo:
            ldrType: null
            geographicArea: "{}"
            qosFulfilInd: "{}"
            twanId: twanId
            ageOfLocationInfo: "{}"
            enodeBId: enodeBId
            plmnId: plmnId
            cellId: cellId
            routingAreaId: routingAreaId
            positionMethod: "{}"
            ueVelocity: "{}"
            trackingAreaId: trackingAreaId
          dddStatus: null
          maxWaitTime: null
          externalId: null
          maxUEAvailabilityTime: null
          failureCause:
            causeType: 9
            s1ApCause: 4
            smCause: 7
            ranapCause: 2
            gmmCause: 3
            bssgpCause: 7
            ranNasCause: ranNasCause
          idleStatusInfo:
            periodicAUTimer: null
            edrxCycleLength: 0.14658129
            activeTime: null
            suggestedNumberOfDlPackets: 0
            idleStatusTimestamp: null
          monitoringType: null
          lossOfConnectReason: 5
          uePerLocationReport:
            externalIds:
            - null
            - null
            msisdns:
            - null
            - null
            ueCount: 0
          roamingStatus: true
          eventTime: null
          dddTrafDescriptor: null
          apiCaps:
          - apiName: apiName
            suppFeat: null
          - apiName: apiName
            suppFeat: null
          pdnConnInfo:
            pdnType: ""
            interfaceInd: ""
            ipv6Addrs:
            - null
            - null
            ipv4Addr: null
            apn: apn
            status: ""
          imeiChange: null
          plmnId: "{}"
          msisdn: null
          reachabilityType: null
        supportedFeatures: "{}"
        maximumResponseTime: null
        locationType: ""
        accuracy: ""
        websockNotifConfig: "{}"
        linearDistance: "{}"
        requestTestNotification: true
        reportingLocEstInd: true
        samplingInterval: null
        svcId: "{}"
        maximumDetectionTime: null
        locationArea5G: "{}"
        maxAgeOfLocEst: "{}"
        velocityRequested: "{}"
        ipv4Addr: "{}"
        msisdn: "{}"
        locationArea: "{}"
        reachabilityType: ""
        dddTraDescriptors:
        - "{}"
        - "{}"
        ldrType: "{}"
        minimumReportInterval: null
        maximumNumberOfReports: 1
        monitorExpireTime: "{}"
        addExtGroupId:
        - null
        - null
        idleStatusIndication: true
        associationType: ""
        externalId: "{}"
        plmnIndication: true
        externalGroupId: "{}"
        repPeriod: "{}"
        mtcProviderId: mtcProviderId
        maxRptExpireIntvl: null
        ipv6Addr: "{}"
        locTimeWindow: "{}"
        monitoringType: ""
        suggestedNumberOfDlPackets: 0
        self: "{}"
        groupReportGuardTime: null
        maximumLatency: null
        apiNames:
        - apiNames
        - apiNames
      anyOf:
      - required:
        - maximumNumberOfReports
      - required:
        - monitorExpireTime
    MonitoringNotification:
      required:
      - subscription
      type: object
      properties:
        subscription:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/Link
        configResults:
          minItems: 1
          type: array
          description: Each element identifies a notification of grouping configuration
            result.
          items:
            $ref: ./TS29122_CommonData.yaml#/components/schemas/ConfigResult
        monitoringEventReports:
          minItems: 1
          type: array
          description: Monitoring event reports.
          items:
            $ref: '#/components/schemas/MonitoringEventReport'
        cancelInd:
          type: boolean
          description: Indicates whether to request to cancel the corresponding monitoring
            subscription. Set to false or omitted otherwise.
        appliedParam:
          $ref: '#/components/schemas/AppliedParameterConfiguration'
    MonitoringEventReport:
      required:
      - monitoringType
      type: object
      properties:
        imeiChange:
          $ref: '#/components/schemas/AssociationType'
        externalId:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/ExternalId
        idleStatusInfo:
          $ref: '#/components/schemas/IdleStatusInfo'
        locationInfo:
          $ref: '#/components/schemas/LocationInfo'
        lossOfConnectReason:
          type: integer
          description: "If \"monitoringType\" is \"LOSS_OF_CONNECTIVITY\", this parameter\
            \ shall be included if available to identify the reason why loss of connectivity\
            \ is reported. Refer to 3GPPÂ TSÂ 29.336Â [11] SubclauseÂ 8.4.58."
        maxUEAvailabilityTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DateTime
        msisdn:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/Msisdn
        monitoringType:
          $ref: '#/components/schemas/MonitoringType'
        uePerLocationReport:
          $ref: '#/components/schemas/UePerLocationReport'
        plmnId:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/PlmnId
        reachabilityType:
          $ref: '#/components/schemas/ReachabilityType'
        roamingStatus:
          type: boolean
          description: "If \"monitoringType\" is \"ROAMING_STATUS\", this parameter\
            \ shall be set to \"true\" if the UE is on roaming status. Set to false\
            \ or omitted otherwise."
        failureCause:
          $ref: '#/components/schemas/FailureCause'
        eventTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DateTime
        pdnConnInfo:
          $ref: '#/components/schemas/PdnConnectionInformation'
        dddStatus:
          $ref: ./TS29571_CommonData.yaml#/components/schemas/DlDataDeliveryStatus
        dddTrafDescriptor:
          $ref: ./TS29571_CommonData.yaml#/components/schemas/DddTrafficDescriptor
        maxWaitTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DateTime
        apiCaps:
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/ApiCapabilityInfo'
      example:
        locationInfo:
          ldrType: null
          geographicArea: "{}"
          qosFulfilInd: "{}"
          twanId: twanId
          ageOfLocationInfo: "{}"
          enodeBId: enodeBId
          plmnId: plmnId
          cellId: cellId
          routingAreaId: routingAreaId
          positionMethod: "{}"
          ueVelocity: "{}"
          trackingAreaId: trackingAreaId
        dddStatus: null
        maxWaitTime: null
        externalId: null
        maxUEAvailabilityTime: null
        failureCause:
          causeType: 9
          s1ApCause: 4
          smCause: 7
          ranapCause: 2
          gmmCause: 3
          bssgpCause: 7
          ranNasCause: ranNasCause
        idleStatusInfo:
          periodicAUTimer: null
          edrxCycleLength: 0.14658129
          activeTime: null
          suggestedNumberOfDlPackets: 0
          idleStatusTimestamp: null
        monitoringType: null
        lossOfConnectReason: 5
        uePerLocationReport:
          externalIds:
          - null
          - null
          msisdns:
          - null
          - null
          ueCount: 0
        roamingStatus: true
        eventTime: null
        dddTrafDescriptor: null
        apiCaps:
        - apiName: apiName
          suppFeat: null
        - apiName: apiName
          suppFeat: null
        pdnConnInfo:
          pdnType: ""
          interfaceInd: ""
          ipv6Addrs:
          - null
          - null
          ipv4Addr: null
          apn: apn
          status: ""
        imeiChange: null
        plmnId: "{}"
        msisdn: null
        reachabilityType: null
    IdleStatusInfo:
      type: object
      properties:
        activeTime:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
        edrxCycleLength:
          minimum: 0
          type: number
          format: float
        suggestedNumberOfDlPackets:
          minimum: 0
          type: integer
          description: Identifies the number of packets shall be buffered in the serving
            gateway. It shall be present if the idle status indication is requested
            by the SCS/AS with "idleStatusIndication" in the "monitoringEventSubscription"
            sets to "true".
        idleStatusTimestamp:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DateTime
        periodicAUTimer:
          $ref: ./TS29122_CommonData.yaml#/components/schemas/DurationSec
      example:
        periodicAUTimer: null
        edrxCycleLength: 0.14658129
        activeTime: null
        suggestedNumberOfDlPackets: 0
        idleStatusTimestamp: null
...
    ApiCapabilityInfo:
      required:
      - apiName
      - suppFeat
      type: object
      properties:
        apiName:
          type: string
        suppFeat:
          $ref: ./TS29571_CommonData.yaml#/components/schemas/SupportedFeatures
      example:
        apiName: apiName
        suppFeat: null
    MonitoringType:
      description: |
        Possible values are - LOSS_OF_CONNECTIVITY: The SCS/AS requests to be notified when the 3GPP network detects that the UE is no longer reachable for signalling or user plane communication - UE_REACHABILITY: The SCS/AS requests to be notified when the UE becomes reachable for sending either SMS or downlink data to the UE - LOCATION_REPORTING: The SCS/AS requests to be notified of the current location or the last known location of the UE - CHANGE_OF_IMSI_IMEI_ASSOCIATION: The SCS/AS requests to be notified when the association of an ME (IMEI(SV)) that uses a specific subscription (IMSI) is changed - ROAMING_STATUS: The SCS/AS queries the UE's current roaming status and requests to get notified when the status changes - COMMUNICATION_FAILURE: The SCS/AS requests to be notified of communication failure events - AVAILABILITY_AFTER_DDN_FAILURE: The SCS/AS requests to be notified when the UE has become available after a DDN failure - NUMBER_OF_UES_IN_AN_AREA: The SCS/AS requests to be notified the number of UEs in a given geographic area  - PDN_CONNECTIVITY_STATUS: The SCS/AS requests to be notified when the 3GPP network detects that the UEâ€™s PDN connection is set up or torn down - DOWNLINK_DATA_DELIVERY_STATUS: The AF requests to be notified when the 3GPP network detects that the downlink data delivery status is changed. - API_SUPPORT_CAPABILITY: The SCS/AS requests to be notified of the availability of support of service APIs.
      anyOf:
      - type: string
        enum:
        - LOSS_OF_CONNECTIVITY
        - UE_REACHABILITY
        - LOCATION_REPORTING
        - CHANGE_OF_IMSI_IMEI_ASSOCIATION
        - ROAMING_STATUS
        - COMMUNICATION_FAILURE
        - AVAILABILITY_AFTER_DDN_FAILURE
        - NUMBER_OF_UES_IN_AN_AREA
        - PDN_CONNECTIVITY_STATUS
        - DOWNLINK_DATA_DELIVERY_STATUS
        - API_SUPPORT_CAPABILITY
      - type: string
        description: |
          This string provides forward-compatibility with future extensions to the enumeration but is not used to encode content defined in the present version of this API.
    ReachabilityType:
      description: |
        Possible values are - SMS : The SCS/AS requests to be notified when the UE becomes reachable for sending SMS to the UE - DATA: The SCS/AS requests to be notified when the UE becomes reachable for sending downlink data to the UE
      anyOf:
      - type: string
        enum:
        - SMS
        - DATA
      - type: string
        description: |
          This string provides forward-compatibility with future extensions to the enumeration but is not used to encode content defined in the present version of this API.
    LocationType:
      description: |
        Possible values are - CURRENT_LOCATION: The SCS/AS requests to be notified for current location - LAST_KNOWN_LOCATION: The SCS/AS requests to be notified for last known location - CURRENT_OR_LAST_KNOWN_LOCATION: The AF requests the current or last known location - INITIAL_LOCATION: The AF requests the initial location
      anyOf:
      - type: string
        enum:
        - CURRENT_LOCATION
        - LAST_KNOWN_LOCATION
        - CURRENT_OR_LAST_KNOWN_LOCATION
        - INITIAL_LOCATION
      - type: string
        description: |
          This string provides forward-compatibility with future extensions to the enumeration but is not used to encode content defined in the present version of this API.
    AssociationType:
      description: |
        Possible values are - IMEI: The value shall be used when the change of IMSI-IMEI association shall be detected - IMEISV: The value shall be used when the change of IMSI-IMEISV association shall be detected
      anyOf:
      - type: string
        enum:
        - IMEI
        - IMEISV
      - type: string
        description: |
          This string provides forward-compatibility with future extensions to the enumeration but is not used to encode content defined in the present version of this API.
    Accuracy:
      description: |
        Possible values are - CGI_ECGI: The SCS/AS requests to be notified at cell level location accuracy. - ENODEB: The SCS/AS requests to be notified at eNodeB level location accuracy. - TA_RA: The SCS/AS requests to be notified at TA/RA level location accuracy. - PLMN: The SCS/AS requests to be notified at PLMN level location accuracy. - TWAN_ID: The SCS/AS requests to be notified at TWAN identifier level location accuracy. - GEO_AREA: The SCS/AS requests to be notified of the geographical area accuracy.
      anyOf:
      - type: string
        enum:
        - CGI_ECGI
        - ENODEB
        - TA_RA
        - PLMN
        - TWAN_ID
        - GEO_AREA
      - type: string
        description: |
          This string provides forward-compatibility with future extensions to the enumeration but is not used to encode content defined in the present version of this API.
...
  securitySchemes:
    oAuth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: "{tokenUrl}"
          scopes: {}